+++ app/jobs/optimia/technical_health/collect_health_job.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Optimia
+  module TechnicalHealth
+    class CollectHealthJob < ApplicationJob
+      queue_as :scheduled_jobs
+
+      def perform
+        Orchestrator.run!
+      end
+    end
+  end
+end