+++ spec/jobs/optimia/technical_health/collect_health_job_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe Optimia::TechnicalHealth::CollectHealthJob do
+  it 'runs the orchestrator' do
+    expect(Optimia::TechnicalHealth::Orchestrator).to receive(:run!)
+
+    described_class.perform_now
+  end
+end